backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
Django==2.2.4
djangorestframework==3.10.2
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
parso==0.5.1
pep8==1.7.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
